+++ packages/server/src/cli/index.ts
@@ -1,0 +1,12 @@
+import { program } from "commander";
+
+import loadEnv from "../env.js";
+import { generateCommands } from "./generate.js";
+import { runCommands } from "./run.js";
+
+loadEnv();
+
+runCommands(program);
+generateCommands(program.command("generate"));
+
+await program.parseAsync();
